Win32: Unicode environment (incoming)
Convert environment from UTF-16 to UTF-8 on startup.

No changes to getenv() are necessary, as the MSVCRT version is implemented
on top of char **environ.

However, we need a minimal putenv implementation, as putenv / _wputenv from
MSVCRT no longer work, for two reasons:
1. they try to keep environ, _wenviron and the process environment block
in sync, using the default system encoding instead of UTF-8 to convert
between charsets
2. msysgit and MSVCRT use different allocators, memory allocated in git
cannot be freed by the CRT and vice versa

Signed-off-by: Karsten Blees <blees@dcon.de>