Win32: Unicode environment (wrapped, incoming)
Replaces Windows "ANSI" environment APIs with their Unicode equivalent,
adding UTF-8/UTF-16LE conversion as necessary.

Getenv converts into and returns a single static buffer which is
overwritten on each call. While this is POSIX-compliant, we need to enable
UNSAFE_GETENV in the Makefile to make it work with git's usage of getenv.

Signed-off-by: Karsten Blees <blees@dcon.de>