Win32: Unicode environment (wrapped, outgoing)
Pass Unicode environment to child processes.

MSVCRT's char **environ is encoded in the default Windows "ANSI" encoding
and thus cannot be used, change to _wenviron.

As a result, the execve/spawnve functions as well as make_augmented_environ
and friends need to use wchar_t** instead of char** for the environment.

Signed-off-by: Karsten Blees <blees@dcon.de>