Win32: keep the environment sorted
The Windows environment is sorted, keep it that way for O(log n)
environment access.

Change env_compare to compare only the keys, so that it can be used to
find an entry irrespective of the value.

Change lookup_env to binary seach for an entry. Return one's complement of
the insert position if not found (libc's bsearch return's NULL).

Implement getenv using lookup_env.

Change env_setenv to insert new entries at the correct position.

Move qsort from mingw_spawnve_fd to mingw_startup. We still need to sort
on startup to make sure that the environment is sorted according to our
env_compare function (while Win32 / CreateProcess requires the environment
to be sorted, it doesn't exactly specify how).

Signed-off-by: Karsten Blees <blees@dcon.de>