add a hashtable implementation that supports O(1) removal
The existing hashtable implementation (in hash.[ch]) uses open addressing
with linear probing, which means that removal is difficult to implement
with less than O(n) complexity. Additionally, entry chaining in case of
collisions is left to the client code.

Add a hashtable implementation that supports O(1) removal and is slightly
easier to use due to builtin entry chaining.

Signed-off-by: Karsten Blees <blees@dcon.de>