compat/terminal.c: always use the shell to read from the Windows terminal
Accessing the Windows console through the special CONIN$ / CONOUT$ devices
doesn't work properly for non-ASCII usernames an passwords.

It also doesn't work for terminal emulators that hide the native console
window (such as mintty), and 'TERM=xterm*' is not necessarily a reliable
indicator for such terminals.

On Windows, always use the shell to read from the terminal. This works fine
for both MSys1 and MSys2, in native console windows as well as mintty, and
properly supports Unicode. It is not even Windows specific (it just needs a
shell that supports 'read -s', 'echo', 'cat' and '/dev/tty').

Signed-off-by: Karsten Blees <blees@dcon.de>